+++ lib/src/widgets/sdui_row.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/widgets.dart';
+import 'package:flutter_sdui/src/widgets/sdui_widget.dart';
+
+/// Represents a Row widget in SDUI.
+class SduiRow extends SduiWidget {
+  final List<SduiWidget> children;
+  final MainAxisAlignment? mainAxisAlignment;
+  final MainAxisSize? mainAxisSize;
+  final CrossAxisAlignment? crossAxisAlignment;
+  final TextDirection? textDirection;
+  final VerticalDirection? verticalDirection;
+  final TextBaseline? textBaseline;
+
+  SduiRow({
+    required this.children,
+    this.mainAxisAlignment,
+    this.mainAxisSize,
+    this.crossAxisAlignment,
+    this.textDirection,
+    this.verticalDirection,
+    this.textBaseline,
+  });
+
+  @override
+  Widget toFlutterWidget() {
+    return Row(
+      mainAxisAlignment: mainAxisAlignment ?? MainAxisAlignment.start,
+      mainAxisSize: mainAxisSize ?? MainAxisSize.max,
+      crossAxisAlignment: crossAxisAlignment ?? CrossAxisAlignment.center,
+      textDirection: textDirection,
+      verticalDirection: verticalDirection ?? VerticalDirection.down,
+      textBaseline: textBaseline,
+      children: children.map((child) => child.toFlutterWidget()).toList(),
+    );
+  }
+}